--- .github/.github/workflows/panvimdoc.yml
@@ -1,25 +1,0 @@
-name: panvimdoc
-
-on:
-  push:
-    branches: [main]
-    paths:
-      - README.md
-      - .github/workflows/panvimdoc.yml
-
-permissions:
-  contents: write
-
-jobs:
-  docs:
-    runs-on: ubuntu-latest
-    name: pandoc to vimdoc
-    steps:
-      - uses: actions/checkout@v2
-      - uses: kdheepak/panvimdoc@main
-        with:
-          vimdoc: themery
-      - uses: stefanzweifel/git-auto-commit-action@v4
-        with:
-          commit_message: "Auto generate docs"
-          branch: ${{ github.head_ref }}
