--- .github/.gitignore
@@ -1,2 +1,0 @@
-doc/tags
-docs/*.xcf
