--- .github/.luarc.json
@@ -1,3 +1,0 @@
-{
-  "diagnostics.globals": ["vim"]
-}
